+++ change_log.md
@@ -5,6 +5,7 @@
 * Add support for PoS [RPC methods]()
 * Split RPC methods to it's own namespace, currently include: `cfx`, `pos`, `trace`, `txpool`
 * Add support for batch RPC
+* Add method `getNextUsableNonce` to `conflux.cfx`, which will first try to use `txpool_nextNonce` to get a usable nonce, if failed it will fall back to `cfx_getNextNonce`
 
 ## v1.7.2