+++ change_log.md
@@ -8,6 +8,13 @@
 * Browser export class name change from Conflux to `TreeGraph`
 * Add method `getNextUsableNonce` to `conflux.advanced`, which will first try to use `txpool_nextNonce` to get a usable nonce, if failed it will fall back to `cfx_getNextNonce`
 
+### trace updates
+
+* Add a new boolean field `valid` to indicate whether this trace has change state
+* The `internal_transfer_action` type trace has add two new field `fromPocket`, `toPocket`
+
+For detail explanation of the trace updates check [this doc](https://github.com/Conflux-Chain/CIPs/issues/88).
+
 Check [v2.0 changes](./docs/v2.0_changes.md) for change details
 
 ## v1.7.2