+++ change_log.md
@@ -19,6 +19,11 @@ For detail explanation of the trace updates check [this doc](https://github.com/
 
 Check [v2.0 changes](./docs/v2.0_changes.md) for change details
 
+### Provider API
+
+* Add `provider.request` to make provider compliant with [EIP-1193](https://eips.ethereum.org/EIPS/eip-1193) 
+* Switch from `provider.call` to `provider.request` in all SDK RPC methods
+
 ## v1.7.2
 
 * Add `balance` key in the result of `estimateGasAndCollateralAdvance` as the balance of `options.from`.