+++ change_log.md
@@ -1,5 +1,9 @@
 # change log
 
+## v2.0.3
+
+* Change trace `epochNumber`, `transactionPosition` from BigInt to Int
+
 ## v2.0.0
 
 * Add support for PoS RPC methods